Refactors, bugfixes and new features
Refactored SongQueue into its own file, fixed various bugs and improved the current play features
